test(bytes): Consolidate escaped tests
